fix: widen channel index wire identity to 32-bit
- ChannelConstants: u16 -> u32 (ControlChannel, Min/MaxDataChannel, Reserved)
- FrameHeader: 8-byte layout -> 12-byte layout
    bytes 0-3: channel index (u32 big-endian)
    byte  4:   flags
    bytes 5-7: reserved (must be zero)
    bytes 8-11: payload length (u32 big-endian)
- ChannelRegistry: allocation counter widened to long, returns uint
- MuxHandshake reconnect: 23-byte/u16 header -> 25-byte/u32 header,
    per-channel replay entry 10 bytes -> 12 bytes (u32 index + u64 pos)
- WriteChannel.RestampChannelIndex: reads payload length from offset 8
- WriteChannel.SetClosed: release closed/unconfirmed slab when transport
    is down to prevent frame accumulation during disconnected open/close
- All tests and docs updated for new wire format

Fixes #512

+++ src/NetConduit/Internal/ChannelBatchRegistrar.cs
@@ -112,7 +112,7 @@ internal sealed class ChannelBatchRegistrar(
         // for the same id is reused. This is essential for composite transit
         // patterns where the peer's INIT for the inbound id may have arrived
         // before the local batch runs.
-        var committedWrites = new List<(ushort Index, WriteChannel Channel)>(count);
+        var committedWrites = new List<(uint Index, WriteChannel Channel)>(count);
         var committedPendingAccepts = new List<ReadChannel>(count);
         // Per-registration committed-channel handle, for Phase 3 assembly.
         var perRegChannel = new IChannel?[count];
@@ -145,7 +145,7 @@ internal sealed class ChannelBatchRegistrar(
                         return false;
                     }
 
-                    ushort idx = registry.AllocateChannelIndex();
+                    uint idx = registry.AllocateChannelIndex();
                     var wc = new WriteChannel(
                         id,
                         idx,
@@ -256,7 +256,7 @@ internal sealed class ChannelBatchRegistrar(
     }
 
     private void RollbackPartialBatch(
-        List<(ushort Index, WriteChannel Channel)> committedWrites,
+        List<(uint Index, WriteChannel Channel)> committedWrites,
         List<ReadChannel> committedPendingAccepts)
     {
         // Caller holds AcceptLock. Unregister in reverse insertion order; channel
